requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.4
django-allauth==0.44.0
django-crispy-forms==1.10.0
django-model-utils==4.1.1
django-rest-framework==0.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.10
oauthlib==3.1.1
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2018.9
requests==2.25.1
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
webdriver-manager==3.4.2
whitenoise==4.1.2
